Fixed error in range search query ancestry
The bug was causing some ancestors of query points
to be removed prematurely if the ancestor was an
antipole on multiple levels of the tree. This did
not affect the final result, and the run time was
actually improved by the bug due to fewer ancestor
checks (in all the scenarios I've tested, at
least), so ancestor checking remains commented out
because it is simply faster that way. A few
functions were changed to return success or
failure status.